src/main/java/org/example/HttpImageStatusCli.java
package org.example;

import java.util.Scanner;

public class HttpImageStatusCli {
    private final HttpStatusImageDownloader downloader = new HttpStatusImageDownloader();

    public void askStatus() {
        Scanner scanner = new Scanner(System.in);

        System.out.print("Enter HTTP status code: ");
        String input = scanner.nextLine();

        try {
            int code = Integer.parseInt(input);
            String imageUrl = downloader.downloadStatusImage(code);
            System.out.println("Image downloaded successfully. You can view it here: " + imageUrl);
        } catch (NumberFormatException e) {
            System.out.println("Please enter a valid number.");
        } catch (Exception e) {
            System.out.println("Error: " + e.getMessage());
        }
    }

    public static void main(String[] args) {
        new HttpImageStatusCli().askStatus();
    }
}
